feat: 초기화 api (#3)

+++ src/main/java/prography/pingpong_game/user/dto/response/UserResponse.java
@@ -6,7 +6,7 @@ import prography.pingpong_game.user.entity.UserStatus;
 
 public record UserResponse(
         Long id,
-        Integer fakerId,
+        Long fakerId,
         String name,
         String email,
         UserStatus status,
